CFD function dummy
Added function to mimic simulation in CFD. A target value of coordinates is specified that would provide the minimal drag. The function finds the difference between the individual coordinate values and the ideal, and sums them up. The closer to the target, the lower the "result" variable is, and thus more favorable.